tutorial/requirements.txt
Django==4.2.5

djangorestframework==3.14.0
fitz==0.0.1.dev2
numpy==1.24.4
opencv_contrib_python==4.8.0.76
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.76
Pillow==7.0.0
Pillow==10.0.1
python_docx==0.8.11
rembg==2.0.50
reportlab==3.5.34
Requests==2.31.0
